# internal/gen/api/spec/openapi.yaml
openapi: 3.1.0
info:
  title: Judex API
  version: 1.0.0
  summary: Judex 协作平台公共 API（docs/plans/v1/06 契约源）
  description: |
    所有业务路径前缀 /api/v1（本规范路径不含该前缀，由网关/服务器统一挂载）。

    协议规则：JSON 字段 camelCase；UUID 字符串；时间 RFC3339 UTC；版本为正整数。
    成功 envelope `{data, meta:{requestId,serverTime,eventCursor?}}`；命令额外返回 `affected`。
    错误 envelope `{error:{code,message,details?,retryable},requestId}`；客户端按 code+details 处理，不解析 message。
    分页 `?limit=50&cursor=<opaque>`（1-100）；列表 data 为 `{items,nextCursor}`。
    有副作用的业务命令要求 Idempotency-Key 头（UUID）；已有对象写操作要求 expectedVersion；正式决定要求 reviewId/reviewHash。
    Web 会话使用 Cookie+CSRF（sessionCookie）；CLI 使用受限 Bearer grant（bearerGrant）。
    阶段实现状态由 GET /system capabilities 返回；未实现的 operation 返回 501 而不是假成功。
servers:
  - url: /api/v1
    description: Same-origin (SPA or CLI with explicit base URL)
...
security:
  - sessionCookie: []
  - bearerGrant: []
paths:
...
  /projects/{projectId}/proposals:
    parameters:
      - $ref: '#/components/parameters/ProjectIdParam'
    get:
      operationId: listProposals
      summary: 提案分页
      tags:
        - proposals
      security:
        - sessionCookie: []
        - bearerGrant: []
      parameters:
        - $ref: '#/components/parameters/LimitParam'
        - $ref: '#/components/parameters/CursorParam'
        - name: status
          in: query
          schema:
            type: string
            enum:
              - draft
              - pending
              - approved
              - cancelled
              - stale
        - name: kind
          in: query
          schema:
            type: string
      responses:
        '200':
          description: 提案分页。
          content:
            application/json:
              schema:
                type: object
                required:
                  - data
                  - meta
                properties:
                  data:
                    type: object
                    required:
                      - items
                      - nextCursor
                    properties:
                      items:
                        type: array
                        items:
                          $ref: '#/components/schemas/Proposal'
                      nextCursor:
                        type:
                          - string
                          - 'null'
                  meta:
                    $ref: '#/components/schemas/RequestMeta'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
    post:
      operationId: createProposal
      summary: 创建 typed 提案草稿（返回差异与服务端建议审批名单）
      tags:
        - proposals
      security:
        - sessionCookie: []
        - bearerGrant: []
      parameters:
        - $ref: '#/components/parameters/IdempotencyKeyParam'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ProposalDraftRequest'
      responses:
        '201':
          description: 草稿已创建。
          content:
            application/json:
              schema:
                type: object
                required:
                  - data
                  - meta
                properties:
                  data:
                    $ref: '#/components/schemas/Proposal'
                  meta:
                    $ref: '#/components/schemas/CommandMeta'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
  /projects/{projectId}/proposals/{proposalId}/draft:
    parameters:
      - $ref: '#/components/parameters/ProjectIdParam'
      - name: proposalId
        in: path
        required: true
        schema:
          type: string
          format: uuid
    put:
      operationId: updateProposalDraft
      summary: 更新提案草稿（只影响 draft）
      tags:
        - proposals
      security:
        - sessionCookie: []
      parameters:
        - $ref: '#/components/parameters/IdempotencyKeyParam'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - expectedVersion
                - changes
              properties:
                expectedVersion:
                  type: integer
                  minimum: 1
                reason:
                  type: string
                  maxLength: 2000
                changes:
                  type: array
                  minItems: 1
                  items:
                    $ref: '#/components/schemas/ProposalChange'
      responses:
        '200':
          description: 更新后的草稿。
          content:
            application/json:
              schema:
                type: object
                required:
                  - data
                  - meta
                properties:
                  data:
                    $ref: '#/components/schemas/Proposal'
                  meta:
                    $ref: '#/components/schemas/CommandMeta'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
  /projects/{projectId}/proposals/{proposalId}/submit:
    parameters:
      - $ref: '#/components/parameters/ProjectIdParam'
      - name: proposalId
        in: path
        required: true
        schema:
          type: string
          format: uuid
    post:
      operationId: submitProposal
      summary: 提交提案（冻结 payload/名单/manifest；sender 人工确认）
      tags:
        - proposals
      security:
        - sessionCookie: []
      parameters:
        - $ref: '#/components/parameters/IdempotencyKeyParam'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ProposalSubmitRequest'
      responses:
        '200':
          description: 冻结后的审阅。
          content:
            application/json:
              schema:
                type: object
                required:
                  - data
                  - meta
                properties:
                  data:
                    $ref: '#/components/schemas/ProposalReview'
                  meta:
                    $ref: '#/components/schemas/CommandMeta'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '409':
          $ref: '#/components/responses/Conflict'
  /projects/{projectId}/proposals/{proposalId}/review:
    parameters:
      - $ref: '#/components/parameters/ProjectIdParam'
      - name: proposalId
        in: path
        required: true
        schema:
          type: string
          format: uuid
    get:
      operationId: getProposalReview
      summary: 固定审阅版本（payload/manifest/slots/votes/可执行动作）
      tags:
        - proposals
      security:
        - sessionCookie: []
        - bearerGrant: []
      responses:
        '200':
          description: 审阅详情。
          content:
            application/json:
              schema:
                type: object
                required:
                  - data
                  - meta
                properties:
                  data:
                    $ref: '#/components/schemas/ProposalReview'
                  meta:
                    $ref: '#/components/schemas/RequestMeta'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
  /projects/{projectId}/proposals/{proposalId}/decisions:
    parameters:
      - $ref: '#/components/parameters/ProjectIdParam'
      - name: proposalId
        in: path
        required: true
        schema:
          type: string
          format: uuid
    post:
      operationId: decideProposal
      summary: 审批决定（当前职责办理人；同一真人可覆盖多个 slot）
      tags:
        - proposals
      security:
        - sessionCookie: []
      parameters:
        - $ref: '#/components/parameters/IdempotencyKeyParam'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ProposalDecisionRequest'
      responses:
        '200':
          description: 决定后的审阅状态（全部通过时同事务应用变更）。
          content:
            application/json:
              schema:
                type: object
                required:
                  - data
                  - meta
                properties:
                  data:
                    $ref: '#/components/schemas/ProposalReview'
                  meta:
                    $ref: '#/components/schemas/CommandMeta'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '409':
          $ref: '#/components/responses/Conflict'
  /projects/{projectId}/proposals/{proposalId}/delegate-decisions:
    parameters:
      - $ref: '#/components/parameters/ProjectIdParam'
      - name: proposalId
        in: path
        required: true
        schema:
          type: string
          format: uuid
    post:
      operationId: delegateProposalDecision
      summary: 授权代批（只覆盖 pending 位，记录实际 actor）
      tags:
        - proposals
      security:
        - sessionCookie: []
      parameters:
        - $ref: '#/components/parameters/IdempotencyKeyParam'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DelegateDecisionRequest'
      responses:
        '200':
          description: 代批后的审阅状态。
          content:
            application/json:
              schema:
                type: object
                required:
                  - data
                  - meta
                properties:
                  data:
                    $ref: '#/components/schemas/ProposalReview'
                  meta:
                    $ref: '#/components/schemas/CommandMeta'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '409':
          $ref: '#/components/responses/Conflict'
  /projects/{projectId}/proposals/{proposalId}/revisions:
    parameters:
      - $ref: '#/components/parameters/ProjectIdParam'
      - name: proposalId
        in: path
        required: true
        schema:
          type: string
          format: uuid
    post:
      operationId: createProposalRevision
      summary: 基于旧审阅创建修订草稿（旧票不继承）
      tags:
        - proposals
      security:
        - sessionCookie: []
      parameters:
        - $ref: '#/components/parameters/IdempotencyKeyParam'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ProposalRevisionRequest'
      responses:
        '201':
          description: 新草稿。
          content:
            application/json:
              schema:
                type: object
                required:
                  - data
                  - meta
                properties:
                  data:
                    $ref: '#/components/schemas/Proposal'
                  meta:
                    $ref: '#/components/schemas/CommandMeta'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
...
components:
  schemas:
...
    RequestMeta:
      type: object
      required:
        - requestId
        - serverTime
      properties:
        requestId:
          type: string
        serverTime:
          type: string
          format: date-time
        eventCursor:
          type: integer
          format: int64
          description: 当前项目事件序列游标，仅项目资源响应返回。
    AffectedResource:
      type: object
      required:
        - type
        - id
        - version
      properties:
        type:
          type: string
        id:
          type: string
        version:
          type: integer
          minimum: 1
    CommandMeta:
      allOf:
        - $ref: '#/components/schemas/RequestMeta'
        - type: object
          properties:
            affected:
              type: array
              items:
                $ref: '#/components/schemas/AffectedResource'
    ErrorBody:
      type: object
      required:
        - code
        - message
      properties:
        code:
          $ref: '#/components/schemas/ErrorCode'
        message:
          type: string
          description: 兜底文案；客户端按 code+details 映射，不解析句子。
        details:
          description: 结构化错误详情，按 code 不同而不同。
        retryable:
          type: boolean
          default: false
    ErrorCode:
      type: string
      enum:
        - VALIDATION_ERROR
        - UNSUPPORTED_SCHEMA
        - INVALID_REFERENCE
        - EMAIL_IN_USE
        - UNAUTHENTICATED
        - SESSION_EXPIRED
        - GRANT_REVOKED
        - FORBIDDEN
        - HUMAN_CONFIRMATION_REQUIRED
        - NOT_FOUND
        - VERSION_CONFLICT
        - REVIEW_STALE
        - DEPENDENCY_CHANGED
        - INVALID_TRANSITION
        - IDEMPOTENCY_CONFLICT
        - SOURCE_VERSION_CONFLICT
        - EVENT_CURSOR_EXPIRED
        - REQUIREMENT_UNMET
        - PAYLOAD_TOO_LARGE
        - UNSUPPORTED_MEDIA
        - RATE_LIMITED
        - BUDGET_EXHAUSTED
        - DEPENDENCY_UNAVAILABLE
        - MODEL_UNAVAILABLE
        - SANDBOX_UNAVAILABLE
        - INTERNAL_ERROR
    ErrorResponse:
      type: object
      required:
        - error
        - requestId
      properties:
        error:
          $ref: '#/components/schemas/ErrorBody'
        requestId:
          type: string
...
    Proposal:
      type: object
      required:
        - id
        - kind
        - status
        - version
        - createdAt
      properties:
        id:
          type: string
          format: uuid
        kind:
          type: string
          enum:
            - work_arrangement
            - work_change
            - dependency_change
            - material_link
            - topic_creation
            - formal_conclusion
            - handoff_change
            - release_acceptance
        status:
          type: string
          enum:
            - draft
            - pending
            - approved
            - cancelled
            - stale
        currentReviewId:
          type:
            - string
            - 'null'
          format: uuid
        topicId:
          type:
            - string
            - 'null'
          format: uuid
        senderUserId:
          type:
            - string
            - 'null'
          format: uuid
        revision:
          type: integer
        version:
          type: integer
        createdAt:
          type: string
          format: date-time
    ProposalChange:
      type: object
      required:
        - operation
        - targetType
      properties:
        operation:
          type: string
          enum:
            - activate_object
            - reference_task
            - create_plan
            - create_task
            - update_scope
            - set_assignment
            - set_requirements
            - link_material
            - link_topic
            - cancel_plan
            - cancel_task
            - update_position
        targetType:
          type: string
        targetId:
          type:
            - string
            - 'null'
          format: uuid
        clientRef:
          type:
            - string
            - 'null'
          description: 新对象临时引用；服务端分配真实 ID 后在结果中映射。
        expectedVersion:
          type: integer
          description: 已有对象的版本校验；新对象省略。
        fields:
          type: object
          additionalProperties: true
        dependsOn:
          type: array
          items:
            type: string
    ProposalDraftRequest:
      type: object
      required:
        - kind
        - changes
      properties:
        kind:
          type: string
          enum:
            - work_arrangement
            - work_change
            - dependency_change
            - material_link
            - topic_creation
            - formal_conclusion
            - handoff_change
            - release_acceptance
        topicId:
          type: string
          format: uuid
        reason:
          type: string
          maxLength: 2000
        changes:
          type: array
          minItems: 1
          items:
            $ref: '#/components/schemas/ProposalChange'
    ProposalReview:
      type: object
      required:
        - reviewId
        - reviewHash
        - proposalId
        - status
        - slots
        - changes
      properties:
        version:
          type: integer
          minimum: 1
        workflowConstraints:
          type: array
          items:
            type: object
            additionalProperties: true
        reviewId:
          type: string
          format: uuid
        reviewHash:
          type: string
        proposalId:
          type: string
          format: uuid
        status:
          type: string
          enum:
            - draft
            - pending
            - approved
            - cancelled
            - stale
        submittedAt:
          type:
            - string
            - 'null'
          format: date-time
        firstApprovalAt:
          type:
            - string
            - 'null'
          format: date-time
        deadlineAt:
          type:
            - string
            - 'null'
          format: date-time
        changes:
          type: array
          items:
            $ref: '#/components/schemas/ProposalChange'
        slots:
          type: array
          items:
            $ref: '#/components/schemas/ApprovalSlot'
        createdIds:
          type: object
          additionalProperties:
            type: string
            format: uuid
        canAct:
          type: boolean
        unmetRequirements:
          type: array
          items:
            type: object
            additionalProperties: true
    ApprovalSlot:
      type: object
      required:
        - id
        - authorityType
        - authorityId
        - state
      properties:
        currentUserId:
          type:
            - string
            - 'null'
          format: uuid
        bindingVersion:
          type: integer
        canDecide:
          type: boolean
        canDelegate:
          type: boolean
        id:
          type: string
          format: uuid
        authorityType:
          type: string
          enum:
            - identity
            - user
        authorityId:
          type: string
          format: uuid
        displayName:
          type: string
        positionName:
          type:
            - string
            - 'null'
        initialUserId:
          type:
            - string
            - 'null'
          format: uuid
        state:
          type: string
          enum:
            - pending
            - approved
            - rejected
            - timeout
            - delegated
    ProposalSubmitRequest:
      type: object
      required:
        - expectedVersion
        - draftHash
      properties:
        expectedVersion:
          type: integer
          minimum: 1
        draftHash:
          type: string
    ProposalDecisionRequest:
      type: object
      required:
        - reviewId
        - reviewHash
        - expectedVersion
        - decision
        - slotIds
      properties:
        reviewId:
          type: string
          format: uuid
        reviewHash:
          type: string
        expectedVersion:
          type: integer
          minimum: 1
        decision:
          type: string
          enum:
            - approve
            - reject
        slotIds:
          type: array
          minItems: 1
          items:
            type: string
            format: uuid
        actingBindingVersions:
          type: array
          items:
            type: object
            required:
              - identityId
              - bindingVersion
            properties:
              identityId:
                type: string
                format: uuid
              bindingVersion:
                type: integer
        reason:
          type: string
          maxLength: 2000
    DelegateDecisionRequest:
      type: object
      required:
        - reviewId
        - reviewHash
        - slotIds
        - reason
      properties:
        reviewId:
          type: string
          format: uuid
        reviewHash:
          type: string
        slotIds:
          type: array
          minItems: 1
          uniqueItems: true
          items:
            type: string
            format: uuid
        reason:
          type: string
          minLength: 1
          maxLength: 2000
    ProposalRevisionRequest:
      type: object
      required:
        - previousReviewId
        - changes
        - reason
      properties:
        previousReviewId:
          type: string
          format: uuid
        reason:
          type: string
          minLength: 1
          maxLength: 2000
        changes:
          type: array
          minItems: 1
          items:
            $ref: '#/components/schemas/ProposalChange'
...
  parameters:
    LimitParam:
      name: limit
      in: query
      description: 分页大小，范围 1-100，默认 50。
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 50
    CursorParam:
      name: cursor
      in: query
      description: 不透明分页游标，编码范围与过滤条件。
      schema:
        type: string
    IdempotencyKeyParam:
      name: Idempotency-Key
      in: header
      required: true
      description: 客户端生成的 UUID；同键同载荷重试返回原结果，同键异载荷返回 409。
      schema:
        type: string
        format: uuid
    ProjectIdParam:
      name: projectId
      in: path
      required: true
      schema:
        type: string
        format: uuid
  responses:
    BadRequest:
      description: 请求不合法（400/422，VALIDATION_ERROR 等）。
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    Unauthorized:
      description: 未认证（401，UNAUTHENTICATED/SESSION_EXPIRED/GRANT_REVOKED）。
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    Forbidden:
      description: 已认证但无权限（403，FORBIDDEN/HUMAN_CONFIRMATION_REQUIRED）。
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    NotFound:
      description: 资源不存在或对非成员隐藏存在性（404）。
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    Conflict:
      description: 状态/版本/幂等冲突（409，详见 error.code）。
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
...
  securitySchemes:
    sessionCookie:
      type: apiKey
      in: cookie
      name: __Host-judex_session
      description: Web 会话 Cookie（本地 HTTP 开发使用 judex_dev_session）。
    bearerGrant:
      type: http
      scheme: bearer
      description: CLI access token（受限 grant，无人工审批能力）。